requirements.txt
psycopg2==2.9.10
Flask==3.1.0
dotenv==0.9.9
pytz==2025.2
azure-storage-blob==12.25.1
azure-identity==1.21.0
